requirements.txt
streamlit>=1.28.0
elevenlabs>=0.2.26
python-dotenv>=1.0.0
httpx>=0.25.0
pytest>=7.4.0
